Expand on purpose in preamble and clarify phrasing
There wasn't much information at the start of the code, so I expanded
upon why it was made, as well as encouraging people to contribute to it,
with the caveat that it might need to be amended based on mandatory
reporting laws in individual locations.

Otherwise, all I did was fix the heading structure, two typos, and
combined two sentences into one.