Update minor text
This changes some small grammar issues; it's not particularly important,
but still annoying. The next big change to the code of conduct will be
to explicit about the fact pretty much everything applies in class as
well as on the social dance floor, which people didn't feel was entirely
clear with the current iteration.