Add Italian translation, replace redunant text with credits
This commit adds an Italian translation by way of Stefano Caltagirone.
With this change I've also removed the copy of the text that was in the
main repository, as the redundancy means it would need to be edited in
multiple places if ever adjusted in the future.

The aesthetic of the page needs to be improved for readability, which
will be done soon, alongside a navigation bar for various languages -
next will be Irish (IE) then French (FR).