Add 'Pobail Stomp', re-order the list of orgs
Pobail Stomp is alive and kicking at the moment; the previous
contributor put Full Swing at the end of the list instead of
alphabetically, but that could've been avoided if I hadn't simply YOLO
accepted the pull request.

Half of the entries on the list are no longer relevant as they are
defunct, but I'm not sure how to reconcile this - they used it, but are
no longer active. Maybe it would make sense to create an "archive"
section, of sorts?