Move defunt events and organisations to "Archive" section
There are a handful of entries that needed to be moved elsewhere, as
they are no longer using the Code of Conduct - because they no longer
exist. I don't want to scrub their existence away completely, so this
seemed like the best compromise.

Of the remaining ones on the list, I'm not certain all of the events
still exist, so I will update their status at a later date.