Update English CoC, Add Spanish CoC, update README
This change addresses some recent concerns with the code of conduct, and
also adds a Spanish localisation. The README has also been restructured
to account for a changelog (Should the need for significantly more
updates arise), and shifts some inactive events into the archive.

An Irish version of the code of conduct will likely be added soon;
French and German are also on the list. I intend to restructure this
website to make the localisations more discoverable, and may host it as
part of a subdomain of swing.ie instead of its current, awkward github
pages website whose domain isn't intuitive to non-technical users.