feat: Update README, add static PDF files, remove web subfolder
I recently transitioned the repository to the Pobail Stomp organisation.
There is some additional housekeeping to do to restructure and rephrase
the text, as the new GitHub pages are serving the README directly as the
live website. It's not a problem, but the current text is very odd as a
result (in reference to "this repository".